Issue #2: Added rebase functionality
Added the ability to rebase a specific migration to a new set of bases.
This could be used to re-position a migration at the end or whatever
else.

Also fixed a bug I hadn't noticed in pruning where if a migration was
pruned with multiple down_migrations, any nodes that had the deleted
migration as a down_migration would not get the full set of
down_migrations

Added entry point executable and tets for the new stuff